split queues.batchSendOptions
