+++ cloudflare/queues/producer_opts.go
@@ -36,31 +36,3 @@ func WithDelay(d time.Duration) SendOption {
 		o.DelaySeconds = int(d.Seconds())
 	}
 }
-
-type batchSendOptions struct {
-	// DelaySeconds - The number of seconds to delay the message.
-	// Default is 0
-	DelaySeconds int
-}
-
-func (o *batchSendOptions) toJS() js.Value {
-	if o == nil {
-		return js.Undefined()
-	}
-
-	obj := jsutil.NewObject()
-	if o.DelaySeconds != 0 {
-		obj.Set("delaySeconds", o.DelaySeconds)
-	}
-
-	return obj
-}
-
-type BatchSendOption func(*batchSendOptions)
-
-// WithBatchDelay changes the number of seconds to delay the message.
-func WithBatchDelay(d time.Duration) BatchSendOption {
-	return func(o *batchSendOptions) {
-		o.DelaySeconds = int(d.Seconds())
-	}
-}
